Move method populate to Helper
